Increase speed of variance calculation
Change data type of gradients from cell to numerical sparse arrays when it results in large speedups (i.e., when having to access them many times).

Bugfix when _I and _F leaf positions are equal (and lie on a bixel boundary).

Change binning of XCAT phantoms to only consider inhaling phases. Since XCAT is symmetrical/there is no hysteresis, this is not a problem.