Changes to motion model evaluation: decimation
Change how the motion model is evaluated. Specifically, the testing data is no longer decimated, but kept at the original sampling frequency. The training data (and subsequent model), on the other hand, are decimated to the desired resampled frequency.

Also add support for varying the resampling frequency as an optimization parameter.